chatbotPython/requirements.txt
transformers==4.35.2
torch==2.0.0
numpy==1.24.3
scikit-learn==1.3.0
sentence-transformers==2.2.2
huggingface-hub==0.16.4
faiss-cpu==1.7.4
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
gunicorn==21.2.0